[Feat-Perf] Use Batching + Squashing  (#5645)
* use folder for slack alerting

* clean up slack alerting

* fix test alerting
