+++ litellm/proxy/proxy_config.yaml
@@ -14,11 +14,9 @@ model_list:
 
 general_settings: 
  master_key: sk-1234 
+ alerting: ["slack"]
+ alerting_threshold: 0.00001
 
 litellm_settings:
-
   callbacks: ["otel"]
-
-  success_callback: ["langsmith", "prometheus"]
-  service_callback: ["prometheus_system"]
